+++ XMPFiles/source/FileHandlers/PSD_Handler.cpp
@@ -1,10 +1,12 @@
 // =================================================================================================
-// ADOBE SYSTEMS INCORPORATED
-// Copyright 2006 Adobe Systems Incorporated
+// Copyright Adobe
+// Copyright 2006 Adobe
 // All Rights Reserved
 //
 // NOTICE: Adobe permits you to use, modify, and distribute this file in accordance with the terms
-// of the Adobe license agreement accompanying it.
+// of the Adobe license agreement accompanying it. If you have received this file from a source other 
+// than Adobe, then your use, modification, or distribution of it requires the prior written permission
+// of Adobe.
 // =================================================================================================
 
 #include "public/include/XMP_Environment.h"	// ! This must be the first include.
@@ -310,7 +312,7 @@ void PSD_MetaHandler::UpdateFile ( bool doSafeUpdate )
 
 		XMP_Assert ( this->xmpPacket.size() == (size_t)oldPacketLength );	// ! Done by common PutXMP logic.
 
-		if ( progressTracker != 0 ) progressTracker->BeginWork ( this->xmpPacket.size() );
+		if ( progressTracker != 0 ) progressTracker->BeginWork ( ((float)(this->xmpPacket.size())) );
 		liveFile->Seek ( oldPacketOffset, kXMP_SeekFromStart  );
 		liveFile->Write ( this->xmpPacket.c_str(), (XMP_StringLen)this->xmpPacket.size() );
 		if ( progressTracker != 0 ) progressTracker->WorkComplete();
